+++ newIDE/app/src/MainFrame/Preferences/PreferencesContext.js
@@ -277,7 +277,7 @@ export type Preferences = {|
   setIsAlwaysOnTopInPreview: (enabled: boolean) => void,
   setEventsSheetCancelInlineParameter: (value: string) => void,
   setShowCommunityExtensions: (enabled: boolean) => void,
-  setShowGetStartedSection: (enabled: boolean) => void,
+  setShowGetStartedSectionByDefault: (enabled: boolean) => void,
   setShowEventBasedObjectsEditor: (enabled: boolean) => void,
   getShowEventBasedObjectsEditor: () => boolean,
   setShowDeprecatedInstructionWarning: (enabled: boolean) => void,
@@ -398,7 +398,7 @@ export const initialPreferences = {
   setIsAlwaysOnTopInPreview: () => {},
   setEventsSheetCancelInlineParameter: () => {},
   setShowCommunityExtensions: () => {},
-  setShowGetStartedSection: (enabled: boolean) => {},
+  setShowGetStartedSectionByDefault: (enabled: boolean) => {},
   setShowEventBasedObjectsEditor: (enabled: boolean) => {},
   getShowEventBasedObjectsEditor: () => false,
   setShowDeprecatedInstructionWarning: (enabled: boolean) => {},
