+++ newIDE/app/src/MainFrame/Preferences/PreferencesProvider.js
@@ -141,7 +141,7 @@ export default class PreferencesProvider extends React.Component<Props, State> {
       this
     ),
     setShowCommunityExtensions: this._setShowCommunityExtensions.bind(this),
-    setShowGetStartedSection: this._setShowGetStartedSection.bind(this),
+    setShowGetStartedSectionByDefault: this._setShowGetStartedSection.bind(this),
     setShowEventBasedObjectsEditor: this._setShowEventBasedObjectsEditor.bind(
       this
     ),
